Handle unreachable code separately
Summary:
This diff changes the toplevels code in order to handle both checking in arbitrary order while still handling unreachable code in execution order.

We may want to reorder checking of statements in arbitrary ways, but that won't affect the program's behavior at runtime, and that includes whether code is reachable or not. Therefore we don't want the reordering of statements to affect which `EUnreachable` errors we provide. This diff ensures that, but recording the *lexically* earliest abnormal control flow encountered and labeling everything afterwards as unreachable, even if typechecking occurred in a different order.

This has some potential performance costs, though they're unlikely to be significant in practice. Previously, unreachable code wasn't typechecked, but now we don't necessarily know whether a statement is reachable or not at the time that we check it, so instead we check everything. However, hopefully unreachable code is not a common occurrence.

Reviewed By: panagosg7

Differential Revision: D27730301

fbshipit-source-id: f2c6584f7b372522bcbb63f0736f0a8166ef2294