Don't key fix cache by reasons
Summary:
When `this`-instantiating a class, we cache the results of these instantiations to ensure we don't perform unnecessary substitutions redundantly. Currently, this cache is keyed by both the instance type and the reason associated with the instantiation, using the logic that the same type instantiated in the same way will always have the same result.

However, `this`-instantiation is special among polymorphic instantiation in that it always has the same argument: every `this`-instantiation of a given class will result in the same instance. So it is unnecessary to differentiate these instantiations by their reason; as things stand right now we are creating a separate instance type for every single instantiation of a class, which is enormously redundant.

This diff keys the fix cache by only the instance type, along with a bit denoting whether the instantiation is an internal or external view of the class, since this will change the argument that the instantiation uses.

Reviewed By: panagosg7

Differential Revision: D27969426

fbshipit-source-id: 2d12f06743dc9371edd509978fbd42ffbc2ce08f