fix mappers that used jsx_identifier
Summary:
all of these `jsx_identifier` mappers were treating the identifier as a variable use, like `<Foo />`. but it is also called for both `foo` and `bar` in `<foo:bar />` and `<foo.bar />`, and in the future will also be called for attribute names.

instead, override `jsx_element_name_identifier` which _only_ matches `<Foo />`.

Reviewed By: samwgoldman

Differential Revision: D28120637

fbshipit-source-id: 727bdb79e568a6246446fc79a8a7d2e76a8784ef