[EnvBuilder] Include locations in refinement info
Summary:
This diff adds the locations of refinements into the refinements generated in the env builder, and adds some tests. Refinements include location sets, and `And`/`Or` refinements union the locations of their operands.

(I realized a little late that this may be stepping on some similar work happening in D27775784 which is unpublished at time of writing; if this is equivalent to/subsumed by that work, this can be abandoned!)

Reviewed By: jbrown215

Differential Revision: D27843371

fbshipit-source-id: 707c243429b7bfa3dc9f909c39fb08a19db0fc38