[newenv] Add def info to provider analysis
Summary: In order to easily go from a def site (any LHS of an assignment) to the provider that governs that variable, we need to track in the entries of the provider analysis what the defs are during the analysis. Then, when we want to find the providers of a def, we look through all the entries to find one where that includes the def, and return the providers for that entry.

Reviewed By: panagosg7

Differential Revision: D27865776

fbshipit-source-id: 2cf3771029ec9ced2beb7d2b28438b00efc79bca