[EnvBuilder] combine refinement and ssi info for specific refiner info
Summary: This diff adds a function to the env builder (which can probably be factored into a future env API module) that determines the locations of both writes and refinements that reach a given read. This (plus handling of general uses, and when combined with other changes to the env builder to make it aware of other aspects of the reordering work) will drive the reordering process, ensuring that specific uses happen after their refiners.

Reviewed By: jbrown215

Differential Revision: D27843467

fbshipit-source-id: b750002ffe9a0dc61b5f18582b48c92bd267f285