[watchman] remove 1 week timeout when waiting for events
Summary:
`Watchman.get_changes` blocks until watchman sends a subscription update, if you set a `~deadline`. If you don't set a deadline, it just checks to see if there are any changes ready right now.

We were always setting a deadline of a week (!!!). A timeout that long is pointless; it might as well wait forever.  Based on our telemetry, this never happens. We'll find out if watchman closes the connection, so we don't have to worry about being stuck forever.

Reviewed By: nmote

Differential Revision: D28021282

fbshipit-source-id: 7f82d3b13690ac77ac6aa31fd84d64875a8e884b