[watchman] make get_sockname return a result
Summary:
this starts to convert all of the functions in watchman.ml to return results instead of throwing exceptions, so that we can ensure we handle all the error cases.

during the transition, we will call `raise_error` to maintain exception behavior, but eventually thread the `result` further and further up and eliminate `try/with`'s.

Reviewed By: nmote

Differential Revision: D28161404

fbshipit-source-id: 62cee5468a4250b9b96082675682160f4a616d9e