Fix precedence issue of NotT
Summary:
Flow_js handles logical operators before some important catch-all cases, such as
`TypeAppT`. This results in having the catch-all case of `NotT` fire on flows of the form
```
TypeAppT _ ~> NotT _
```
thus missing the chance to instantiate the LHS:
```
type A<X> = X;
declare var a: A<false>;
const a1: false = !a; // no error
const a2: true = !a; // error
```

This diff moves handling of the unary `NotT` past the wildcards:
```
TypeAppT _, _
UnionT _, _
```
to avoid situations like the above.

With this change we can also omit the special case of `UnionT` with `NotT`.

Reviewed By: dsainati1

Differential Revision: D28242267

fbshipit-source-id: 7f89a1338aefafcea67d7e80964b7cef1dc24960