implement Smart Select (textDocument/selectionRange)
Summary:
`textDocument/selectionRange` takes one or more cursor positions and returns a tree of ranges that should be selected as you expand or shrink the selection. in vscode, you can do this via ctrl+shift+cmd+(left or right arrow).

for example, you might expand from an identifier to the containing expression, to the whole statement, to the whole block, to the whole method, to the whole class, to the whole file.

Reviewed By: gkz

Differential Revision: D28093731

fbshipit-source-id: 326e0c4abbb8a75d4f60dcafdbb8a187b9041031