Update saved_state_skip_lazy_recheck test to ES6 imports/exports
Summary:
This is in preparation for D28338781, which turns on `new_check` mode for this test. The new type sig hashing is conservative with CommonJS modules, so if the hash of a module's exports change, the hash of a type imported from that module will also change.

This is because type exports do not have their own namespace, and we can't be sure without firing up the typechecker that the value assigned to `module.exports` isn't an object with a property that should be type-importable, such as a class. Because the whole point of new check mode is to avoid running the typechecker in this step, we have to be conservative when hashing.

Fortunately, ES6 imports/exports are much easier to statically analyze, so they do not have the same pitfall. After this change, the test passes in D28338781 when `new_check` mode is turned on without any further modifications.

Reviewed By: samwgoldman

Differential Revision: D28339239

fbshipit-source-id: b80cc1df4075955299b9cad560f3e088efe0b965