fail faster when watchman restarts
Summary:
when watchman disconnects, we try to reconnect. we previously first re-watched the project and *then* asked if watchman had restarted. watching the project can take a long time -- especially if watchman restarted! and we exit if it did, so that time was wasted.

instead, ask watchman if it restarted first.

if the root is not watched yet, I *think* watchman may sometimes error[1], but if the root isn't watched anymore then we missed updates and need to restart, so that error is actually fine for this case.

[1] the query command hits [this code](https://github.com/facebook/watchman/blob/ea2444f3aec858cc23f061a3ce61c84f7616b8a4/cmds/query.cpp#L15), which ends up [here](https://github.com/facebook/watchman/blob/db731a47e217b39e414bb2e0f254887c9d17aadc/root/resolve.cpp#L167-L170) with `auto_watch = false`, and I think that means it can error with "directory X is not watched"

Reviewed By: nmote

Differential Revision: D28276188

fbshipit-source-id: f95c735c1996658b2b213c19160b43784809564d