speed up has_watchman_restarted_since by up to 60s
Summary:
when our watchman connection dies, we send a new query to see if watchman restarted; if so, we have to restart too since we may have missed updates.

this query will wait for watchman to synchronize (finish watching the folder and processing changes). we don't care what the actual state of the filesystem is, only whether it's a new instance of watchman, so setting `sync_timeout` to 0 (disabling it) makes it return instantaneously.

before, it could take up to 60 seconds. if it timed out, that would cause us to exit -- either incorrectly, because the sync was just slow, or for the wrong reason (timeout vs restart).

Reviewed By: nmote

Differential Revision: D28273185

fbshipit-source-id: abd94fc25a693e35aa533422321b786a89752990