[EnvBuilder] Havoc only when variable is considered havocable
Summary:
This diff uses the invalidation api to only havoc when a variable is determined as being havocable by the invalidation_api (the same as is used to determine if it's havocable during checking). In order to use the invalidation api, we need to have already run an SSA analysis over the program, so this adds another pass: first we run the SSA builder on our program, then the env builder.

One change that needs to be made here is that `havoc_current_ssa_env` needs to now know if the havocing is happening as a result of a call or of a yield, because more things need to be havoced by yields. This distinction is now introduced, but unused, in the SSA builder itself; it becomes relevant in the env builder.

We don't need to change `havoc_uninitialized_ssa_env`, which is called when entering a new var scope, because we *want* to unconditionally havoc everything when we enter a function, since we don't know from where it's being called.

Reviewed By: jbrown215

Differential Revision: D27871038

fbshipit-source-id: b5e071764690ffc4d3bf1208e8b52e920de9d764