[enums] Fix enum declaration visitor
Summary:
The `flow_ast_mapper` method for visiting enum declarations originally visited the name of the enum using the `identifier` method. Typically, `identifier` is used for visiting *reads* of expressions, but the identifier in an enum declaration is a *write*, which should be visited using `pattern_identifier` (see corresponding pattern in ex. class and function declarations). `pattern_identifier` takes an optional `kind`, which represents what kind of variable is being declared (when this is a declaration and not just a write); in the case of enums I think this is `const`, since enums compile to `const` declarations (right?)

This change fixes the `ssa_builder` test case for enums without any other change needing to be made because the ssa builder now correctly understands the enum identifier as a write rather than a read.

Reviewed By: gkz

Differential Revision: D28204020

fbshipit-source-id: e9b596beaa465b70c1bbc2ca2cdd21bec0881ec3