[envbuilder] When the EnvBuilder havocs, use providers
Summary:
This diff essentially implements the general/specific logic for havocing without needing to explicitly represent things as general or specific. Previously, when the Env builder havoced, it merged the ssa entry's val ref with a special havoc val ref, which accumulates all writes to the variable; so essentially when havocing happened we'd add all writes to the variable to the set of live writes that reach any later reads.

This diff changes that behavior for the env builder in order to match the logic used for building the environment, rather than matching the actual behavior of the program. Specifically, we consider that after havocing all variables are in a General state, where any specific information has been lost. In a general state, variables are constrained by their providers rather than any control-flow based information from the ssa builder.

To represent this logic in the env builder, we now, instead of merging the current val ref with the havoc val ref, we just use the havoc val ref on its own. And, instead of adding all writes to the havoc val ref, the havoc val ref will specifically include only the providers for the variable, according to the provider analysis.

This means that the env builder now needs to know about the provider analysis. Rather than passing provider info into the env builder, this diff has the `program` entry point for the env builder run the provider analysis directly, and include the provider output as part of the result of running the env builder.

Reviewed By: jbrown215

Differential Revision: D28080163

fbshipit-source-id: f94bd2fdf4096112726573009a418a39396364a9