[providers] Synthesize entry for undeclared variables in the provider analysis
Summary:
This diff modifies the provider_finder to handle undeclared variables, e.g. `x = 42` when `x` is undeclared. Obviously this is an error, but we should handle it anyways. This diff changes the process of adding a def location (and potentially a provider location) to an entry to, if the entry wasn't created during the declaration-finding pass (as is the case for an undeclared variable), create a new entry for the variable in the closest var scope to the declaration.

This process needs to explore all scopes outwards in order to make sure that the variable doesn't already exist in an outer scope, so it passes a lazy tuple containing the new entry and the function for updating it recursively, and only forces it if the variable is never found.

As a minor update this changes the `find_entries_for_existing_variable` to `find_entry_for_existing_variable`: instead of finding the set of entries for the scope that contains the searched-for variable, it just returns the entry itself.

Reviewed By: panagosg7

Differential Revision: D28204116

fbshipit-source-id: 3e1fc4346acb0c02ab2dfd90121c95a0fb8d7944