[newenv] Determine ordering using providers and envbuilder
Summary:
This diff starts to build `order_builder`, the module that will be used to determine the order in which programs are checked.

It operates by, for each statement in a list, visiting all the identifier nodes and adding the providers or refiners they rely on (represented as locations) to a set. For RHS identifiers, this means looking into the env builder and finding the writes and refinements that reach the variable. For LHS identifiers, first we check if the location is a provider itself, in which case we do nothing--a provider for a variable doesn't depend on already knowing what the type of the variable is (pretty much by definition). Otherwise, e.g. when it's an assignment that isn't also a provider, we return the set of providers for that variable, because writes to variables are constrained by the types of the variable's providers.

In order to handle block statements that include both a write to a variable, and a read specific to that write (e.g. `if (...) { x = 42; console.log(x) }`), when we finish processing a statement we remove from the set all locations that are included within the set's location.

This analysis gives us a mapping from statements to the provider/refiner locations that the statement depends on. We can then figure out which statement those locations are included in by checking loc inclusion, which ultimately gives us a mapping from statements to statements.

Reviewed By: panagosg7

Differential Revision: D28039883

fbshipit-source-id: e6f56d296bb38a0c519ff44b7ecfa8fc473c570d