[enums] Eliminate `enum_name` from enum type and use reasons instead
Summary:
The enum type contains `enum_name`, a string, which is the name of the enum at the declaration site. We don't need this and can use the reason of the type instead (also improves some errors by pointing to the source location referenced).

This change is required if we want to support generic enum types, since such a type doesn't derive from a enum declaration in the code that would give it a name. This also unlocks allowing anonymous enums when used with `export default` (`e.g. `export default enum {A, B}`), as is allowed with function declarations and classes, if we want to add such a feature.

Reviewed By: panagosg7

Differential Revision: D28104551

fbshipit-source-id: 196f6adc447ad43e41a1fafc8a78ca3944fca44a