refactor handling of missing include paths
Summary:
watchman only watches one root folder, but filters results to subpaths. for each `[include]` path, watchman splits it into the `watch_root` and the relative path. if the path doesn't exist, we try to figure out the relative path it if it's a child of the root.

we collect the watch_root returned for each include. then we check whether the missing paths are children of any of those roots. then, if there's more than one root, we error.

this diff reverses the checks, so we error first if there's more than one root; then we can more simply check whether the missing includes are children of that root.

Reviewed By: nmote

Differential Revision: D28236903

fbshipit-source-id: d900f62315f9029f1010b603c43e6238a66b74ac