make Watchman.transform_asynchronous_get_changes_response return a result
Summary:
this stack converts all of the functions in watchman.ml to return results instead of throwing exceptions, so that we can ensure we handle all the error cases.

during the transition, we will call `raise_error` to maintain exception behavior, but eventually thread the `result` further and further up and eliminate `try/with`'s.

Reviewed By: nmote

Differential Revision: D28173780

fbshipit-source-id: ff8d3228214bcf4810d6dea2bb507a63f49ed99f