Implement recheck optimizations in new check mode
Summary:
Recheck optimizations let Flow skip work which will certainly have no meaningful
effects. Specifically, for a file in the fanout of a given change, if none of
its dependencies have meaningfully changed, we don't need to actually check that
file.

(This optimization is the same as "early cutoff" described in Build Systems a la
Carte)

Before new check mode, we would detect changes by computing a hash of the
signature context. This hash is computed by visiting the context starting at
exports, accumulating interesting bits of information as we do the traversal.

In new check mode, there are no signature contexts. Instead, we compute a hash
of the type signature of the file. This is equivalent by simple reasoning:
signature contexts are the output of a deterministic function from type
signatures. The output (signature context) can only change if the input (type
signatures) also change.

Happily, hashing type signatures is also significantly more efficient compared
to hashing signature contexts. Because sharing is explicit in type signatures,
we can avoid repeatedly re-hashing the same data, which we currently do when
hashing signature contexts.

Each object in a type signature which might be reachable from multiple paths
(e.g., local defs) is converted to a node in a potentially cyclic graph. We use
the Cycle_hash module (D27538588) to visit this graph and propagate changes.

For a given object, we compute the initial hash by simply hashing the serialized
OCaml data for that signature. This is a conservative approximation, which should
capture any significant changes to the file itself. As an intuition for this,
consider that any information which is not present in the signature could not
possibly be significant to dependent files.

Reviewed By: nmote

Differential Revision: D28036916

fbshipit-source-id: 65ee27c15f1da8142d50aa47a2232eb5ff469856