Deserialize type sigs lazily during check
Summary:
Before this diff, when creating the Type_sig_merge.file record for a dependency
during check, we would read out the entire type sig for the dependency from
shared memory and deserialize it into OCaml values.

This diff makes this process lazier, so we only read and deserialize the
signature data that we actually need during check.

Reviewed By: panagosg7

Differential Revision: D28012183

fbshipit-source-id: 3bef6fa29c4b95d7881fc575886ab977e8c3511d