Represent empty addr tbls as null
Summary:
Before this diff, an empty addr tbl was allocated in the shared heap as a single
word with Addr_tbl_tag and size 0. Instead, we can represent the address of an
empty addr tbl as the null addr. This means fewer objects in the heap and less
work for GC.

Reviewed By: mroch

Differential Revision: D25488426

fbshipit-source-id: 18339a316872910f19d6e27db9f818a40b57c302