Store exports in heap as separate objects
Summary:
Before this change, a checked file's exports were stored as a single serialized
OCaml value, plus an optional "export def" object, which stored the
`module.exports` or default export for CJS and ES modules, respectively.

This change instead stores each each type export, named export (ES), and
the module.exports (CJS) as a separate heap object, as well as a "module info"
object which stores the names of the exported objects as well as export *
information and whether a module is strict.

A checked file can have an ES or CJS module type. We differentiate in the heap
by inspecting the tag in the header. The OCaml API to the heap makes this type
safe by introducing a "dyn_module" phantom type for the typed addr scheme used
by SharedMem.NewAPI. We can construct a dyn_module from a cjs_module/es_module
and we can read out a dyn_module into either a cjs_module or es_module.

By separating the exports like this, we can give each export its own signature
hash in shared memory. Later diffs in the stack will add a new field to these
objects which stores a hash, which we can then use to granularly propagate
changes for recheck optimizations.

Reviewed By: nmote

Differential Revision: D28015712

fbshipit-source-id: 2074bc09e3b616be7acffe48a9d26fdde8d1fdb9