Add space for signature hashes in new heap
Summary:
This diff adds a new field to the heap objects representing exports, where we
will store signature hashes. We store a hash for each export as well as a hash
for the module itself.

When computing the hashes for a dependent file, we will read from these fields
for each dependency. If we can resolve an import to a specific export, we will
only incorporate that export's hash.

For now, we only store hashes with exports to save space. We could also store
hashes alongside each reachable definition, but there's no current use case for
that.

Reviewed By: panagosg7

Differential Revision: D28030604

fbshipit-source-id: bdf6719300455f785ee340e62f49d2e9dfc7dd9b