Migrate saved_state_skip_lazy_recheck test to new check mode
Summary:
All the tests are run with saved state turned on, and then separately with new check mode turned on. But, they aren't run with both options turned on, and this test is configured to only run with saved state, so it wasn't running in new check mode previously.

Instead of addressing this generally, especially because new check mode will become the only mode relatively soon, I just turn it on for this test.

Reviewed By: samwgoldman

Differential Revision: D28338781

fbshipit-source-id: 6ec5449bc38ef4f6b285ba0c3dc0f33fa64e90d0