Use root-relative paths when sighashing in new check mode
Summary:
This unlocks the ability to use sighashes loaded from saved state to skip work in the check phase. I did much of the work to enable this in D25137486 (https://github.com/facebook/flow/commit/1c218b08441a7d5dcb8ffc874856d18adb303638), but was left with two blockers related to hash stability:

1. The hashes were often unstable between Flow runs. In practice, I found that ~40% of files in a large real-world codebase had sig hashes that were unstable in this way. Within a single Flow run, they were stable, but this instability between runs meant that we couldn't make effective use of these sighashes loaded from saved state, since they were computed in a previous run. Because we cannot skip checking a file if ANY dependency has changed, this meant that in practice we were only able to skip a very small fraction of files.
2. As part of the sighash, we hash the path to modules that are referenced. Flow stores all paths internally as absolute paths. This means that the hashes will all be different if they are loaded from a saved state that was generated in a copy of the repository with a different absolute path. This makes them nearly useless in practice.

For issue (1), I decided that instead of spending time debugging the existing, complicated sighashes, I would just wait for types-first 2.0. In types-first 2.0, signatures are now a tailor-made data structure rather than large blobs of types dumped from the type system. I bet that this principled approach would lead to more stable hashes, and my bet paid off: some initial testing indicates that less than half a percent of files have sighashes that are perturbed between runs. It should be zero, but this is still good enough that we will see most of the performance benefits that we can get from this work. In the future, it would most likely be worth addressing the remaining hash instability.

For issue (2), I attempted to relativize paths upon sighashing in D25432952, but despite my best efforts I was not able to prevent a significant merge time regression. Even avoiding all allocations when relativizing by implementing the relativization via string prefix matching in C, this additional work led to a 9% merge time regression.

I considered two courses of action to address this:
* Change Flow to use relative paths throughout. This would be a lot of work, but would likely also lead to many other benefits. For example, we spend a lot of time starting up from saved state just making the relative paths stored in saved state absolute again for use in the rest of Flow. Using relative paths everywhere would significantly improve startup time. However, once I started looking into this I realized that it would be a very large project, and decided not to take it on at the moment.
* Wait for types-first 2.0, since it dramatically improves the whole situation around signatures and hashing.

I ended up going with the second option, so here we are. Signatures and signature hashing is now a much more principled process, and one of the many changes is that we hash paths much less frequently so this isn't a particularly hot path anymore.

`Files.relative_path` is horribly inefficient, so we might expect this change to significantly affect merge time if it's called enough. Fortunately, that doesn't appear to be the case. Types-first 2.0 has already dramatically reduced merge time so that it's not a substantial portion of total time spent, and this change appears to increase it by only 2.5%.

Reviewed By: samwgoldman

Differential Revision: D28294663

fbshipit-source-id: 969d2e3cf5564bc292adaee6d35b9d2b505d4b85