[enums] Fix implicit void return analysis when exhaustively checked switch in switch
Summary:
If you had a switch which checked every case of a Flow Enum, but did not return in the final case and did not have a break in the final case, we inaccurately reported that there was no implicit void return.

E.g. this should error (if you add a `break` it does error).
```
enum Foo {
  A,
}
function f(x: Foo): string {
  switch (x) {
    case Foo.A:
  }
}
```

This fix results in a small change in behaviour: if you have a switch of a Flow Enum nested in a `case` of a switch of a Flow Enum, and you're returning from every branch, you must add a `break` now. E.g.

```
enum E {A, B}
function f(x: E, y: E) {
  switch (x) {
    case E.A:
      return 1;
    case E.B:
      switch (y) {
        case E.A:
          return 2;
        case E.B:
          return 3;
      }
      break; // <-- This `break` is now required.
  }
}
```
This isn't a big deal since ESLint already complains if you miss adding a `break` to any `case` (other than the last one), so this actually makes the behaviour more consistent.

If in the future I think of a way to support it without the added break then I'll do so.

Reviewed By: dsainati1

Differential Revision: D28318244

fbshipit-source-id: f0f8abde1f6646c79c4ad863132f5465b96af7ba