Remove unnecessary trigger character checks from autocomplete
Summary: The type inference hooks used in autocomplete were made to return false if the autocomplete request came from a trigger character, back in a time when the hooks were what told autocomplete whether to service an autocomplete request or not. Nowadays we usually always want to respond to an autocomplete request if we can, and the only thing the type inference hooks do is that they tell the type inference algorithm to ignore malformed code at the autocomplete cursor position, and to set things up a certain way for autocomplete. These hooks being guarded on the trigger character was causing object key autocomplete to fail in some cases.

Reviewed By: mroch

Differential Revision: D28445415

fbshipit-source-id: 738359dcbb086418bac673db2327f5a9a8d9c002