Change typed autocomplete to compute Type.t upper bound instead of Ty.t upper bound
Summary: Same logic as before, but computes the upper bound as a Type.t instead of as a Ty.t. This is useful because for object key autocomplete we want to reuse this code for computing an upper bound, but in that case the computed upper bound will need to be something we can pass to `members_of_type` which takes a Type.t.

Reviewed By: panagosg7

Differential Revision: D28445413

fbshipit-source-id: 51620c44ac05911dd9f564140dc2ec23539db571