Make new check file cache bounded in size
Summary:
Before this change, the file cache and ccx cache grew unbounded. This is not
enturely unreasonable when called from workers, for two reasons: (1) workers
process a fixed number of files before exiting, and (2) workers have a memory
cap, so even if the caches get too large before finishing a batch, the worker
will still exit.

However, Merge_js.merge_contents is called from the server, which is a long-
lived process. Currently each time we call merge_contents we create fresh file
and ccx caches. However, it would be more efficient to be able to re-use caches
across calls to merge_contents.

Because the server process is long-lived, we need some bound on the size of
these caches. This diff implements a LRU cache with a fixed bound.

Currently we still create fresh caches for each call of merge_contents. This API
enables sharing the cache between invocations and invalidating the caches as
needed, but that is currently unused.

Reviewed By: nmote

Differential Revision: D28487667

fbshipit-source-id: 83bf52d7706fefc68266d8fb0668b1a5ff2c2823