Reuse file cache across calls to check_context_contents
Summary:
Under new check mode, IDE features that rely on merge_contents regressed. The
caching in this diff should address the regression by reducing the amount of
work performed on each call.

Before, we created the necessary types for any dependencies every time we called
merge_contents. In practice it is very likely that we will type check file
contents having a very similar set of dependencies, so this sharing is very
effective.

The cache is bounded, so it's okay to keep the cache around, although we might
consider clearing the cache after some amount of idling to reduce memory use of
the server process, but this is left for future work.

It is important, however, that we clear out any changed files from the cache. If
a dependency changes, the cached types might no longer be accurate. In "old
check" mode, where we use the SigContextHeap to get dependency types, this would
happen transparently**, but we need to be a bit more explicit.

Happily, the set of new and changed files, which is determined by signature
hashing during merge, corresponds to the entries which need to be removed. That
is, if a file's signature hash is changed, then it's exports have changed, and
any cached type information need to be recalculated.

** Note that the SharedMem.WithCache API actually does not cache old entries!
Whenever we call SigContextHeap.get_old we are deserializing data from shared
memory every time.

Reviewed By: nmote

Differential Revision: D28500124

fbshipit-source-id: 321eacfdb23a11ac842f5c2c847f9149dfdb4100