retry on error during watchman init
Summary:
previously, when we first try to initialize watchman and it errors for any reason, we give up and exit. this can lead to `Out of retries` errors if watchman takes long enough to come back alive.

say for example, watchman is starting up and its socket isn't ready, which happens frequently because we restart when it restarts. there's a bit of a race where our new server might try to connect before the new watchman is ready.

now, we retry the init up to 8 times (the arbitrary limit we use for retries if we get an error after init), with a delay of 4, 8, 16, 32, 32, 32, 32, and 32 seconds between each retry (total of about 3 minutes).

Reviewed By: samwgoldman

Differential Revision: D28432570

fbshipit-source-id: 4eb90ec01e91c08182047ce931f87d3911ac4bde