remove 'phase parameter from constraints
Summary:
I had created this distinction when the idea of having a mode where *only*
"fully-resolved" type variables would be used was appealing.

As it turns out, even in the new "annotation-only" mode, we still need an
"unresolved" state, to represent (for example) intermediate versions of
recursive types.

While it's cool that the type system allows us to make this distinction it doesn't
look like I'll be using it in for the moment, so I'm backing out this change to
keep the codebase a bit cleaner.

Instead, my plan is to use a separate "annotation" graph with a different data
type of constraints.

Reviewed By: samwgoldman

Differential Revision: D28491502

fbshipit-source-id: e7f5b615d377210d9285a8a0883e7e1c56b0f020