error on class-object subtyping
Summary:
This restructures the class-object-interface hierarchy. Previously, objects and interfaces were equivalent, and both were valid supertypes of a class that had matching properties. This change reorders this relationship to put interfaces at the top, allowing neither objects nor classes to supertype an interface. Objects are also unable to supertype classes anymore. Interfaces remain supertypes of both objects and classes.

Changes in tests show some examples of differences in how interface and object types will need to be used after this change.

This also updates the definition of `$ArrayLike` in our core libraries to be an interface, since this is required for it to be a supertype of the `Array` and `$ReadOnlyArray` builtin classes. One side effect of this is it makes strings valid subtypes of `$ArrayLike`, which makes sense since they have a length property, can be indexed, and are iterable. Tests are updated to reflect this behavior.

Reviewed By: mvitousek

Differential Revision: D25982688

fbshipit-source-id: a1801ea5fed83bb704f632164c86f6104d81ca1c