[envbuilder] Handle globals in env builder
Summary: In order to deal with globals in the env builder, which don't have writes to them and so don't otherwise appear in the SSA env (but which *can* be refined), this diff adds a new `globals_env` map to the env_builder. This maps from names to ids drawn from the same set used in `ssa` values, and these ids serve as keys to the refinement map. When we try to refine a variable that doesn't exist in the ssa_env, we add that variable to the globals map instead. When we havoc, we just reset the globals map entirely.

Reviewed By: panagosg7

Differential Revision: D28204243

fbshipit-source-id: 3f3ef30e404f9bcd522b8dc351d33b02b4a13b95