[envbuilder] Refactor envbuilder output type
Summary:
The env_builder's output tuple was getting a little hairy, lol. This diff replaces the tuple with a record.

Of note is that the record also includes the regular ssa_builder's `ssa_values`. This is to support D28204287, which will use the env_builder rather than the ssa_builder during "real" typechecking, but which still needs to include the ssa_values in order to drive the current implementation of refinement invalidation.

Reviewed By: panagosg7

Differential Revision: D28204288

fbshipit-source-id: df963c93c3930f244ae8e0c79878797bfa7e5960