[new-env] Add cacheing to invalidation api
Summary: Previously, whenever the env builder havoced, the invalidation api would fully compute whether or not each variable in scope was constlike and/or captured by closures. This information is constant per-variable, so there's no need to recompute this every time. This diff adds caches that records this information for each variable, which greatly improves performance of the env builder.

Reviewed By: panagosg7

Differential Revision: D28241708

fbshipit-source-id: 40d9215dfe026dc1f041289acf327384c16ba536